removed warnings & reformatted classHeader.js
